Move #fget from Cache to ClientAPI
